+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Check for [NameStartChar](https://www.w3.org/TR/xml/#NT-NameStartChar)
   during qualified name parsing.
 
+### Removed
+- `StreamError::InvalidAttributeValue`. `StreamError::InvalidChar` will be emitted instead.
+
 ## [0.9.0] - 2019-02-27
 ### Added
 - `span` field to all `Token` variants, which contains a whole token span in bytes.